Use babel to minify instead of webpack's uglify.js default
Makes for more robust target code. Skip mangling to allow for inspector to work properly (anyway, it's better if the code is understandable). Size is affect a bit (double the size, hmm). We could counter this by gzipping on server. I hope it doesn't affect runtime performance too much on slow devices.