Add inventory state tracking
In LBA2, inventory items are stored using game vars. But, in addition to
that, there is an array of "inventory state". This is used to distinguish
between variants of the same item. In the game, there are the following
state-dependent items:
 - tunic vs wizard tunic
 - pearl vs travel token (this one is weird!)
 - laser pistol vs empty pistol vs crystal fragment
 - proto pack vs jet pack
 - blowgun vs blowtron
 - Francos fragment vs Dark Monk key

As a drive-by fix, remove all hard-coded '40' values for the LBA2
inventory size. And fix the incorrect SET_USED_INVENTORY opcode.