Update item data now that we have inventory state
With the ability to track the game's inventory state from the previous
commit, we can now distinuish beween magic ball colours, laser pistol
state, blowgun/blowtron state, etc. This commit updates the inventory
data to support these differences.

This has a cascading effect on other code. We now need to distinguish
between a number of different ways of identifying inventory contents:
 - slot number: this is the index of the slot where the item is shown in
   inventory UI. All laser pistol items have the same slot number.
 - raw item number: this is the index of the game var which is used to
   track whether Twinsen has the item. This value is the same for all
   variants of the same item. All laser pistol items have the same raw
   item number.
 - item number: unique identifier for the item variant. Same as the raw
   item number for items without variants. All laser pistol items have
   different item numbers.
 - resource id: the index within the game resources of the model and
   text for the item variant. All laser pistol items have different
   resource IDs.

The rules for where the different types should be used are:
 - slot: only used by the Inventory UI
 - raw item id: item identity (same as game var ID). Visible to scripts.
 - item id: item appearance, used to select Twinsen's model
 - resource id: used by inventory UI to get the right model and name