Slightly improve sprite rendering
This commit adds some additional cases for how we turn sprites into 3D
objects: as well as the existing XY and ZY plane meshes, this adds XZ
plane meshes and a fallback isometric-aligned mesh.

Some objects are still drawn poorly (particularly in the fallback case)
but they tend to be better than before.