Implement loading (but not animation) of Anim3DS
The ANIM3DS.HQR file stores these sprites as a series of frames and an
info block giving the frame range of each animation. The sprites
themselves are normal sprites.

This change ensures that actors with the Anim3DS flag load the right
sprites but they are not actually animated at this point. For now, we
just render the first sprite in the sequence. To see an example, visit
the Sendell's Ball room in the sewers.