Fix clock resets when pausing
When restarting the clock, its elapsed time gets set to zero. Because we
track waits according to the elapsed time, this causes all wait ops in life
and move scripts to wait for too long after the game is paused (which
stops and restarts the clock).

To avoid this, track the total elapsed time every time the clock is
stopped.