Implement 'low collisions' mode for block collisions
Based on testing LBA2, low collisions mode (enabled when the brick
collision mode is set to 2; 0 is disabled and 1 is enabled) is used when
Twinsen crawls through a low passage. Without this mode, he gets stuck
on the scenery.

This change arbitrarily halves the height of Twinsen's collision box in
this mode, which seems to be enough to get him through.

A good place to test this is the passage between the weather wizard's
tent and the sewers.