Implement interleaved switch/if statements
The LBA games make use of this unusual construct to make certain switch
cases only apply if a surrounding 'if' passes. To support them in text
mode, a bunch of custom indent code can be removed and it works
afterwards.

The fix for Blockliy is more complicated - essentially, arbitrary life
script commands are now allowed inside a switch statement. Extra care
then has to be taken as the switch statement is not neccessarily the
direct parent block of the case blocks.