Implement Blockly rich dropdowns in case statements
Unlike if statements, for cases, the context isn't immediately available.
This is fine for most condition types but with any types that rely on
having an actor value (anims, bodies and tracks) or variable IDs
(var_values), this was lost so only an unannotated dropdown was used.

With this change, the information is now passed down from the switch to
the cases so they have the appropriate context available.