Extracted twinkling star shader
Attempt to extract and adapt the twinkling star shader from the Dome level. Added two of those around Mr. Kurtz's golden globe in scene 173 (entry 117).

Fix: Reversed vertex index order so the front side of the triangles would be drawn facing the camera. That way, the star doesn't need to be double sided ✌️